refactor(functions): split index.ts into focused modules
Organize Cloud Functions into alerts, subscriptions, config, firebase,
and shared utils for easier maintenance without behavior changes.
