+++ functions/src/firebase.ts
@@ -1,0 +1,6 @@
+import { initializeApp } from "firebase-admin/app";
+import { getFirestore } from "firebase-admin/firestore";
+
+initializeApp();
+
+export const db = getFirestore();
